# api/openapi.yaml
openapi: 3.0.3
info:
  title: TradingCopilot API
  version: 0.1.0
  description: JSON:API contract target for the Go modular monolith migration.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /wake-plans:
    get:
      operationId: getWakePlans
      parameters:
        - $ref: "#/components/parameters/ResearchTeamId"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/MeetingIdQuery"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/WakePlanCollectionDocument" }
    post:
      operationId: postWakePlans
      requestBody: { $ref: "#/components/requestBodies/WakePlanUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/WakePlanDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
...
  /wake-plans/{planId}:
    put:
      operationId: putWakePlan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody: { $ref: "#/components/requestBodies/WakePlanUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/WakePlanDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
        "404": { $ref: "#/components/responses/JsonApiErrorDocument" }
    delete:
      operationId: deleteWakePlan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200": { $ref: "#/components/responses/DeletionDocument" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    WakePlanUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanUpsertDocument"
...
  responses:
...
    JsonApiErrorDocument:
      description: JSON:API error document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/JsonApiErrorDocument"
...
    DeletionDocument:
      description: Deletion JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/DeletionDocument"
    WakePlanCollectionDocument:
      description: Wake plan JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanCollectionDocument"
    WakePlanDocument:
      description: Wake plan JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanDocument"
...
  parameters:
...
    Status:
      name: status
      in: query
      schema: { type: string }
...
    MeetingIdQuery:
      name: meetingId
      in: query
      schema: { type: string }
    ResearchTeamId:
      name: researchTeamId
      in: query
      schema: { type: string }
...
    PageLimit:
      name: page[limit]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    PageCursor:
      name: page[cursor]
      in: query
      schema:
        type: string
...
    PlanId: { name: planId, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    JsonApiErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ErrorObject"
...
    DeletionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/DeletionResource"
    DeletionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [deletions]
        id:
          type: string
        attributes:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [deleted]
            unboundAccountIds:
              type: array
              items:
                type: integer
            unboundAccountCount:
              type: integer
    WakePlanCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WakePlanResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    WakePlanDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/WakePlanResource"
    WakePlanResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [wake-plans]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/WakePlanAttributes"
        relationships:
          type: object
          properties:
            meeting:
              $ref: "#/components/schemas/Relationship"
            sourceMeetingEvent:
              $ref: "#/components/schemas/Relationship"
    WakePlanAttributes:
      type: object
      required: [researchTeamId, triggerType, triggerConfig, reason, status, createdAt]
      properties:
        researchTeamId: { type: integer }
        meetingId:
          type: integer
          nullable: true
        triggerType:
          type: string
        triggerConfig:
          type: object
          additionalProperties: true
        reason:
          type: string
        sourceMeetingEventId:
          type: integer
          nullable: true
        sourceRoleKey:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paused, fired, cancelled]
        nextCheckAt:
          type: string
          format: date-time
          nullable: true
        firedAt:
          type: string
          format: date-time
          nullable: true
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        resultSummary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    WakePlanUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type:
              type: string
              enum: [wake-plans]
            id:
              type: string
            attributes:
              $ref: "#/components/schemas/WakePlanInputAttributes"
    WakePlanInputAttributes:
      type: object
      required: [researchTeamId, triggerType, triggerConfig, reason]
      properties:
        researchTeamId: { type: integer }
        meetingId:
          type: integer
          nullable: true
        triggerType:
          type: string
          enum: [time, indicator, event]
        triggerConfig:
          type: object
          additionalProperties: true
        reason:
          type: string
        sourceMeetingEventId:
          type: integer
          nullable: true
        sourceRoleKey:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paused, fired, cancelled]
        nextCheckAt:
          type: string
          format: date-time
          nullable: true
...
    Relationship:
      type: object
      properties:
        data:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ResourceIdentifier"
            - type: array
              items:
                $ref: "#/components/schemas/ResourceIdentifier"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    ResourceIdentifier:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
        id:
          type: string
        meta:
          type: object
          additionalProperties: true
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
        field:
          type: string
    Links:
      type: object
      properties:
        self:
          type: string
        next:
          type: string
        prev:
          type: string